pyproject.toml
[tool.poetry]
name = "tep"
version = "0.1.0"
description = ""
authors = "Mohsen Seyedkazemi Ardebili"
email = "mohsen.seyedkazemi@unibo.it"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
requests = "^2.32.3"
black = "^24.8.0"
pylint = "^3.3.1"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
